physics 클래스명 대문자로 RigidCircle로 변경 -> 알고리즘 코드 부분 분리 중
