+++ cpp_pj.vcxproj.filters
@@ -27,6 +27,9 @@
     <ClCompile Include="OperateGL.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="CheckWinner.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="DrawView.h">
@@ -38,10 +41,13 @@
     <ClInclude Include="OperateGL.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="physics.h">
+    <ClInclude Include="RigidCircle.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="main.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CheckWinner.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
