+++ main.h
@@ -1,4 +1,4 @@
-#include "physics.h"
+#include "RigidCircle.h"
 
 void check_over(int xPos, int yPos, int color);
 void over();
